split msg ID type and enum definition
